+++ CHANGELOG.rst
@@ -7,13 +7,14 @@
 Changelog
 =========
 
-1.2.0 - 2022.xx.xx
+1.2.0 - 2022.10.21
 ------------------
 
 **New features**
 
 - Implemented specification of number of counterexamples in :meth:`~datajudge.WithinRequirement.add_varchar_regex_constraint`.
 - Implemented in-database regex matching for some dialects via ``computation_in_db`` parameter in :meth:`~datajudge.WithinRequirement.add_varchar_regex_constraint`.
+- Added support for BigQuery backends.
 
 
 **Bug fix:**